Add table display for library
A single 'updateLibraryDisplay' function is used to handle displaying
data since it means after modifying the library array, calling
updateLibraryDisplay ensures the html exactly matches it.

Otherwise, parity between html and array is more complicated. After
modifying the array, you must find the relevant html and modify it in
the same way, effectively doubling your code for every modification.

While re-generating large sets of data is inefficient, this project is
so small the performance impact is not a concern and the simpler code is
worth the tradeoff.