Refactor books to use id for search-related functions
Add a read-only id field to the book class and use it for lookups in the
library array, instead of using a book's array index. This is specified
by the update in The Odin Project's instructions for the project:
https://www.theodinproject.com/lessons/node-path-javascript-library

While this project's scope/functionality is small enough that using the
array index of books as an "id" is not yet an issue, it causes many
problems long-term.

For instance, if you store a books' array index in the program and then
that book moves positions in the array, using that index to access the
book will yield a different book than your original target.

It is a better practice to give each book a unique, unchanging
identifier instead of relying on a books' index in the array.